pyproject.toml
[tool.poetry]
name = "he_scheduling"
version = "0.5.0"
description = "A FastAPI scheduling microservice"
authors = ["Michiel Jansen <michiel.mj@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/michielmj/he-scheduling"
repository = "https://github.com/michielmj/he-scheduling"
documentation = "https://github.com/michielmj/he-scheduling"
keywords = ["fastapi", "microservice", "api", "web", "scheduling"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Framework :: FastAPI",
]
packages = [{ include = "he_scheduling" }]

# Specify the entry point for your application
[tool.poetry.scripts]
start-fastapi-server = "he_scheduling.main:run"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
fastapi = "^0.115.0"
uvicorn = "^0.30.0"
ortools = "^9.11.4210"
pydantic = "^2.0.0"
celery = "^5.4.0"
pydantic-settings = "^2.6.1"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
flower = "^2.0.1"


[tool.poetry.dev-dependencies]
pytest = "^8.3.3"
black = "^24.8.0"
flake8 = "^7.1.1"
httpx = "^0.27.2"
mypy = "^1.11.2"
setuptools = "^75.1.0"

[tool.poetry.group.dev.dependencies]
pytest-mock = "^3.14.0"
pika = "^1.3.2"

[tool.pytest.ini_options]
testpaths = ["tests"]  # Specify where pytest should look for tests
addopts = "--disable-warnings"  # Additional options for pytest

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
